requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.3
django-allauth==0.33.0
django-datatable==0.3.1
django-formtools==2.0
django-rest-auth==0.9.1
djangorestframework==3.6.3
idna==2.6
mysqlclient==1.3.4
oauthlib==2.0.4
olefile==0.44
Pillow==4.2.1
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
django-cors-headers==2.1.0

